requirements.txt
# DEEPRUST - Local Deepfake Detection CLI

# Core Inference
transformers==4.46.3
torch==2.5.1
huggingface-hub==0.26.2
Pillow==11.0.0
librosa==0.10.2.post1
soundfile==0.12.1
numpy==1.26.4

# Utils
python-dotenv==1.0.1
tqdm==4.66.5
colorama==0.4.6
